Repo: Add a shell script to check cross-repository links
One huge issue with software documentation is that it's often outdated. In order to make it easier to synchronize code changes with the documentation pages that refer to them, I'll try to put in place an automated alert.

This is how I imagine it working:

1) Documentation contains code snippets with URL to the code itself (in another repo) [TODO]
2) Shell script collects these links and ensures they're still valid [this commit]
3) The script runs periodically and on each new commit, ensuring that desyncs notify me [TODO]

This is a proof of concept and will likely require refinement.