Theme: Swizzle the builtin MDXComponents component
This is an awkward way of getting to auto-import the custom components that are embedded in the markdown files... but there's no better way of doing this apparently (wrapping this component is "forbidden", pff).

I don't plan on customizing anything here really, so hopefully the maintenance overhead for updating the index will be limited.